Still allow both max auditors and max mediators.

+++ dart-common/src/lib.rs
@@ -23,7 +23,7 @@ pub const MAX_ASSET_ID: u32 = ((1_u64 << ASSET_ID_BITS) - 1) as u32;
 pub const MAX_CURVE_TREE_GENS: usize = 1 << 13;
 pub const MAX_ASSET_AUDITORS: u32 = 2;
 pub const MAX_ASSET_MEDIATORS: u32 = 2;
-pub const MAX_ASSET_KEYS: usize = (MAX_ASSET_AUDITORS + MAX_ASSET_MEDIATORS) as usize;
+pub const MAX_ASSET_KEYS: u32 = MAX_ASSET_AUDITORS + MAX_ASSET_MEDIATORS;
 
 pub const ACCOUNT_TREE_L: usize = 64;
 pub const ACCOUNT_TREE_M: usize = 1;
